separate function for reviews pagination
